feat: search anywhere for posts & style: arrange username and prompt

+++ src/Controller/DeleteController.php
@@ -4,6 +4,7 @@ namespace App\Controller;
 
 use App\Entity\Post;
 use App\Repository\UserRepository;
+use App\Service\SecurityManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -19,13 +20,19 @@ class DeleteController extends AbstractController
         Post                   $post,
         EntityManagerInterface $entityManager,
         UserRepository         $userRepository,
+        SecurityManager        $securityManager
     ): Response
     {
         $owner = $userRepository->findOneBy(['username' => $this->getUser()->getUserIdentifier()]);
 
-        if ($this->isCsrfTokenValid('delete' . $post->getId(), $request->request->get('_token'))) {
-            $entityManager->remove($post);
-            $entityManager->flush();
+        if ($securityManager->isOwnerOfPost($owner, $post)) {
+            if ($this->isCsrfTokenValid('delete' . $post->getId(), $request->request->get('_token'))) {
+                $entityManager->remove($post);
+                $entityManager->flush();
+                $this->addFlash('success', 'Post supprimé !');
+            }
+        } else {
+            $this->addFlash('danger', 'T\'est pas propriétaire du post fréro');
         }
 
         return $this->redirectToRoute('app_user_page', ['username' => $owner->getUserIdentifier()], Response::HTTP_SEE_OTHER);
